組織のリポジトリも100以内に限り読み込ませることができるようになった
100超える場合も対応したいなぁ面倒だなぁ

+++ ReadGitHub.sh
@@ -1,19 +1,35 @@
 #!/bin/bash
-if [ $# != 1 ]; then
-    echo 引数エラー: $*
-    exit 1
-else
-    echo OK
+for OPT in "$@"
+do
+    case $OPT in
+        -u)
+            USER=1
+            NAME=$2
+            ;;
+        -o)
+            ORG=1
+            NAME=$2
+            shift
+            ;;
+    esac
+    shift
+done
+if [ $USER ];then
+    echo aa
+    my-repositories $NAME >> ./$NAME.list
+elif [ $ORG ];then
+    curl -u mirachan010 "https://api.github.com/orgs/$NAME/repos?per_page=100&page=1">$NAME_1
+    cat $NAME_1|jq -r .[].name > $NAME.list
+    echo `cat $NAME.list|wc -l`
 fi
-my-repositories $1 > ./$1.list
-Max=`cat $1.list |wc -l`
+Max=`cat $NAME.list |wc -l`
 COUNT=0
 VCOUNT=0
 while read line
 do
-    git clone git://github.com/$1/$line --depth=1
+    git clone git://github.com/$NAME/$line --depth=1
     if [ -d ./$line/autoload -o -d ./$line/plugin ];then
-        echo $1/$line >> list_new
+        echo $NAME/$line >> list_new
         VCOUNT=$(( VCOUNT + 1 ))
     fi
     COUNT=$(( COUNT + 1 ))
@@ -21,10 +37,12 @@ do
     rm -rf ./$line &
     sleep 10 &
     wait
-done < $1.list
+done < $NAME.list
 echo checking
-comm <(sort ./list_new) <(sort ./list) -23 >> list
-rm ./$1.list
-rm ./list_new
+if [ -f ./list_new ];then
+    comm <(sort ./list_new) <(sort ./list) -23 >> list
+    rm ./list_new
+fi
+rm ./$NAME.list
 sort -u list -o list
 echo finish
